% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cramer.R
\name{cvm}
\alias{cvm}
\title{function:two-sample Cramer-von Mises test for equal distributions}
\usage{
cvm(x, y, data)
}
\arguments{
\item{x}{a vector of data}

\item{y}{a vector of data}

\item{data}{data sources}
}
\value{
Histogram of  data
}
\description{
function:two-sample Cramer-von Mises test for equal distributions
}
\examples{
\dontrun{
attach(chickwts)
x <- sort(as.vector(weight[feed == "soybean"]))
y <- sort(as.vector(weight[feed == "linseed"]))
cvm1 <- cvm(x,y,"Example 8.1")
}
}
