% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Norm.R
\name{Norm}
\alias{Norm}
\title{A Metropolis-Hastings sampler using R}
\usage{
Norm(a, b, N)
}
\arguments{
\item{a}{Mean Value of Normal Distribution}

\item{b}{Variance of Normal Distribution}

\item{N}{Test number}
}
\value{
Histogram of Sampling data
}
\description{
A Metropolis-Hastings sampler using R
}
\examples{
\dontrun{
Norm(3,5,10000)
}
}
