% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QQPLOT.R
\name{normtest}
\alias{normtest}
\title{A test of normality using R}
\usage{
normtest(data, alpha = 0.05)
}
\arguments{
\item{data}{A vector of data}

\item{alpha}{Threshold of test}
}
\value{
QQ-plot, histogram and test result of shapiro.test
}
\description{
A test of normality using R
}
\examples{
\dontrun{
data = rnorm(3000)
normtest(data)
}
}
